Building with --release works now (use e.g., cargo make build-unity)
The problem:
build-dependencies are always build in dev, even if you specify --release. So cs_generator always rebuilt rust_bindings without optimizations. Additionally it seems, when using --release the dev profile specified in the toml is ignored, but when not using --release the dev profile is not ignored.

Fix:
in Cargo.toml we specified than for --release we always use optimizations for the build-dependencies